feat(Change account system and subscriptions, updating, reports, components, dashboards and commands)

+++ database/migrations/2024_10_12_123208_create_announcements_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->string('announce_title');
             $table->text('description');
             $table->dateTime('expiration_time');
-            $table->bigInteger('salary');
+            $table->decimal('salary', 10, 0);
             $table->boolean('pro')->default(false);
             $table->unsignedBigInteger('area_id');
             $table->foreignUuid('user_id');
